requirements.txt
osgeo.ogr
psycopg2
